Add files via upload
# Classification Cancer Study Case
In this code I will analyse which is the best classification algorithm to predict a benign/malign cancer tumor. 
You can find the database using the sklearn databases or accessing through the UCI ML Repository. 
Breast Cancer Wisconsin (Diagnostic) datasets.
https://goo.gl/U2Uwz2
The dataset contains 569 observations of patients with 30 numeric attributes and kind of tumor.
Here, I will take only 75% of the sample to train the algorithms and then compare among them which one makes the best predictions.
The algorithms used are:


*   Random Forest Classification
*   Decision Tree Classification 
* Linear SVM
* Kernel SVM
* Naive Bayes
* Logistic Regression